GH-2576: Fix No Seeks After Error Handling
Resolves https://github.com/spring-projects/spring-kafka/issues/2576

When the error handler `seeksAfterError` is false, we keep the unprocessed
records in memory (`retainedRecords`), pause the consumer, and attempt
redelivery.

If a rebalance occurs during this state, the behavior depends on whether
a legacy or cooperative partition assignor is used.

With a legacy assignor, all partitions are unassigned and usually a subset
is reassigned. This causes their positions to be reset to the last committed
offset. In this case, all retained records must be cleared, otherwise an
emergency stop occurs because records are polled when we don't expect them.

With a cooperative assignor, only a subset of the partitions are revoked.
Records for these partitions must be pruned from the `retainedRecords` and
not redelivered to the listener.

- add code to the rebalance listener to prune the revoked paritions from the retained records
- re-pause any assigned partitions if retained records are present

Add tests for both types of assignor.

Also tested with the reporter's reproducer.

**cherry-pick to 2.9.x**

* Fix typo.
# Conflicts:
#	spring-kafka/src/main/java/org/springframework/kafka/listener/KafkaMessageListenerContainer.java