Relativize sourceDir property of asciidoctorPdf
This sets the path sensitivity to relative for the sourceDir property
of the asciidoctorPdf task. Previously, the full absolute path to the
sources directory would be considered as an input to the task. This
would cause a local build cache miss when executing the task from two
different directories or a remote build cache hit given that the paths
to the sources directory on CI and for local developers will undoubtedly
be different.