GH-2817: KafkaTemplate: No double error timers (#2823)
* GH-2817: KafkaTemplate: No double error timers

Fixes https://github.com/spring-projects/spring-kafka/issues/2817

An observation in `KafkaTemplate` can be marked with error from a `Callback`.
Then `Future` is evaluated and its exception is thrown back to the `observeSend()`.
Here this exception is caught and reported to the observation again.
This creates a second timer in the Micrometer, but with different error tag

* Check for error presence in the `observeSend()` `catch` block and skip
second report

**Cherry-pick to `3.0.x`**

* * Fix import order for Checkstyle rule