GH-2991: Fix unwanted warning logs in KMLC
Fixes: #2991

There is a warning log from `KafkaMessageListenerContainter#checkRebalanceCommits`
that unconditionally prints that offsets could not be committed. Because of this
unconditional check, the messages are very likely to be false positives. Fixing
this by adding a check to see if there are indeed uncommitted records.

(cherry picked from commit f9476a380ac442f41ccf6441d7af8f51a9be09a3)