Remove redundant `artifacts` from Gradle config
This also causes the problem with early `api` task initialization
which leads to an attempt to resolve dependencies.
The `io.spring.dependency-management` has not done its job for each module yet in that case.
Anyway these artifacts are now the part of `publishing`, so no need in their definitions in addition.