GH-3176: Take default.dsl.store into account
Fixes: #3176

Currently, the applications need to use special customizers to use in-memory state stores.
Kafka Streams provides a property to configure that - `default.dsl.store` (KIP-591), which the
framework does not take into account when creating the `StreamsBuilder` since it ignores any
config and always relies on customizations after the fact. This change allows this property to
work in the framework so that applications can easily switch to in-memory state stores.

**Auto-cherry-pick to `3.1.x` & `3.0.x`**